requirements/prod.txt
Django==1.11.6
Pillow==4.3.0
django-allauth==0.33.0
django-braces==1.11.0
django-cities-light==3.3.0
django-ckeditor==5.3.1
django-crispy-forms==1.7.0
django-guardian==1.4.9
django-imagekit==4.0.1
djangocms-admin-style==1.2.7
psycopg2==2.7.3.2
python3-openid==3.1.0
